apps/pumpfun-bot/requirements.txt
python-dotenv
websockets
aiohttp
requests
base58
solders
solana
uvloop; platform_system != "Windows"